feat: finished rock paper scissors
It took longer than i expected, and i'm sure there is a better way to do things.
But i'm proud of this one, outside of the random choice from the computer i got everything on my own which was my goal.
I don't know if a while loop is the way to go though. I did this exercise once in python, waaay easier syntax. I was thinking of a switch statement or maybe another type of loop. But i generally like the gameOver = false approach. Excercise done :D.