feat: delete every loop and games references
so, i had to delete everything in the other branches because i was making a mess. So just deleted everything that was cluttering the game and simplified it to it's core. Now i just have basic functions, not even a callout to these functions, but something is more than nothing i guess. The gameover, games left, player scores and winner declaration is gone now. With a couple of tweaks, this game will sure go for eternity.